core/requirements.txt
Django>=4.1,<5.0

python-decouple>=3.5,<4.0 

pillow

djangorestframework==3.15.2
